Only report error message when process returns an error code
ExManCmd annoyingly doesn't use stderr to report so we're stuck fetching error codes out of message with a regex. This, unfortunatelly is prone to false positives in rare cases. When install fails we receive a process code, so we can use that to determine that the install failed, though the code itself doesn't appear to correlate with anything, and doesn't follow the documented error codes for child_process